separate middleware from api
